refactor: extract optional runtime sidecar boundary
